+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next
 - Changed `from_pipelines` to accept an iterator instead of vector
+- Add `Pipeline::from_rows()`
 
 ## 0.3.1 - 2023 Feb 17
 - Fix `Pipeline::from_path` panic